+++ dds/src/implementation/rtps/messages/mod.rs
@@ -1,7 +1,4 @@
-use self::overall_structure::SubmessageHeaderRead;
-
 pub mod overall_structure;
 pub mod submessage_elements;
 pub mod submessages;
 pub mod types;
-pub mod submessage_elements_write;
